Add comments with links, undo change that doesn't work
I have a webcam now, but I don't think I'll have time to test the code
today so I'll probably do that tomorrow. Also, the second link that I
added as a comment is just a better implementation that I will probably
switch to. However, it may or may not sacrifice accuracy (that's
something that needs to be tested)