Initial Passenger-WSGI file
Passenger is pretty weird, so I have no clue whether this will work. On
top of that, Passenger has no mechanism for debugging - if something is
wrong there's a 500 error with no context. Hooray. Still, there's
actually a pretty good chance that this could work, so I'd like to try
it out. This particular setup relies on django so I guess we need to
take that into account. If anyone else understands passenger at all,
feel free to change the file.